backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==0.7.1
pydantic==1.10.22
google-cloud-aiplatform==1.38.1
google-cloud-vision==3.4.4
google-cloud-speech==2.21.0
python-dotenv==1.0.0
gunicorn==21.2.0
uvicorn[standard]==0.24.0
